Move legend data to constants.
