+++ packages/common/src/utils/prepareCookieStatsComponents.ts
@@ -48,7 +48,7 @@ const prepareCookieStatsComponents = (
       count: cookieStats.blockedCookies[reason] ?? 0,
       //@ts-ignore
       color: BLOCKED_COLOR_MAP[reason],
-      countClassName: `text-${reason.toLowerCase()}`,
+      countClassName: 'text-' + reason.toLowerCase(),
     });
   });
